Fix test setup for `generateMetadata` parent handling
Can't expect metadata that reads the pathname (i.e. canonical URL) to be
included in the fallback shell. 🤦🏻‍♂️
