+++ test/e2e/app-dir/use-cache/app/(partially-static)/generate-metadata-resume/layout.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next'
 
-export async function generateMetadata(): Promise<Metadata> {
-  return { metadataBase: new URL('https://example.com/foo') }
+export const metadata: Metadata = {
+  metadataBase: new URL('https://example.com/foo'),
 }
 
 export default function Layout({ children }: { children: React.ReactNode }) {
