+++ test/e2e/app-dir/use-cache/app/(partially-static)/generate-metadata-resume/nested/page.tsx
@@ -1,19 +1,24 @@
-import { Metadata } from 'next'
+import { Metadata, ResolvingMetadata } from 'next'
 import { connection } from 'next/server'
 import { Suspense } from 'react'
 
-export async function generateMetadata({
-  searchParams,
-}: {
-  searchParams: Promise<Record<string, string | string[] | undefined>>
-}): Promise<Metadata> {
+export async function generateMetadata(
+  _: { searchParams: Promise<Record<string, string | string[] | undefined>> },
+  parent: ResolvingMetadata
+): Promise<Metadata> {
   'use cache'
 
   // Explicitly not reading search params here. The search params should be
   // omitted from the cache key, so that we ensure a cache hit when resuming the
   // partially prerendered page.
 
-  return { title: new Date().toISOString() }
+  const { metadataBase } = await parent
+
+  return {
+    title: new Date().toISOString(),
+    metadataBase: metadataBase?.replace('/bar', '/baz'),
+    alternates: { canonical: '/qux' },
+  }
 }
 
 export default function Page() {
