+++ test/e2e/app-dir/use-cache/app/(partially-static)/generate-metadata-resume/params-unused/[slug]/layout.tsx
@@ -1,9 +1,8 @@
-import { Metadata, ResolvingMetadata } from 'next'
+import { Metadata } from 'next'
 
-export async function generateMetadata(
-  { params }: { params: Promise<{ slug: string }> },
-  parent: ResolvingMetadata
-): Promise<Metadata> {
+export async function generateMetadata(_: {
+  params: Promise<{ slug: string }>
+}): Promise<Metadata> {
   'use cache'
 
   // Explicitly not reading params here. The description should appear in the
@@ -11,13 +10,7 @@ export async function generateMetadata(
   // cache hit (from the RDC), but omitting unused params from cache keys (and
   // upgrading cache keys when they are used) is not yet implemented.
 
-  const { metadataBase } = await parent
-
-  return {
-    description: new Date().toISOString(),
-    // We can not return a URL instance from a `'use cache'` function.
-    metadataBase: metadataBase?.replace('/foo', '/bar'),
-  }
+  return { description: new Date().toISOString() }
 }
 
 export default function Layout({ children }: { children: React.ReactNode }) {
