+++ test/e2e/app-dir/use-cache/app/(partially-static)/generate-metadata-resume/params-unused/[slug]/page.tsx
@@ -1,11 +1,10 @@
-import { Metadata, ResolvingMetadata } from 'next'
+import { Metadata } from 'next'
 import { connection } from 'next/server'
 import { Suspense } from 'react'
 
-export async function generateMetadata(
-  { params }: { params: Promise<{ slug: string }> },
-  parent: ResolvingMetadata
-): Promise<Metadata> {
+export async function generateMetadata(_: {
+  params: Promise<{ slug: string }>
+}): Promise<Metadata> {
   'use cache'
 
   // Explicitly not reading params here. The title should appear in the
@@ -13,13 +12,7 @@ export async function generateMetadata(
   // cache hit (from the RDC), but omitting unused params from cache keys (and
   // upgrading cache keys when they are used) is not yet implemented.
 
-  const { metadataBase } = await parent
-
-  return {
-    title: new Date().toISOString(),
-    metadataBase: metadataBase?.replace('/bar', '/baz'),
-    alternates: { canonical: '/qux' },
-  }
+  return { title: new Date().toISOString() }
 }
 
 export default function Page() {
